# Deep copy AST in compile