--- a/src/browser.js
+++ b/src/browser.js
@@ -22,13 +22,13 @@ packages.forEach(function(pkg) {
   pkg.headers.forEach(webppl.requireHeaderWrapper);
 });
 
+var extra = webppl.parsePackageCode(packages);
+
 function run(code, k, verbose) {
-  var extra = webppl.parsePackageCode(packages, verbose);
   return webppl.run(code, k, { extra: extra, verbose: verbose });
 }
 
 function compile(code, verbose) {
-  var extra = webppl.parsePackageCode(packages, verbose);
   return webppl.compile(code, { extra: extra, verbose: verbose });
 }