At present, `compile` mutates package ASTs. (#247) This fixes that by having `compile` copy the ASTs before performing any transforms. I [tried this approach before](https://github.com/probmods/webppl/pull/241#issuecomment-147342288) but found it to be too slow. I realised it can be made to work by writing a simpler deep copy than the fancy one in lodash which I used originally.

With this in place, it's now possible to call `parsePackageCode` once on page load, rather than in every call to `run` and `compile` in the browser. (This PR includes that change too.) Hopefully this will make running code boxes feel a bit snappier. Rough before and after timings are something like:

Before

````
1st run:
parsePackageCode: 700ms
compile: 300ms

Subsequent runs:
parsePackageCode: 200ms
compile: 200ms
````

After:

````
Page load:
parsePackageCode: 700ms

1st run:
compile: 320ms

Subsequent runs:
compile: 200ms
````

So, for all but the first run, the amount of time it takes before the program starts running drops from around 400ms to 200ms with this change. (Ignoring any other steps, non of which I think take much time.) The win might be bigger with extra packages added to the browser bundle.

As things stand, the package/header ASTs are copied when running under Node, even though `compile` is only ever called once there. We could add an option to `compile` to avoid doing this, but copying the header AST only takes about 10ms so I wasn't convinced it was worth the extra complexity. Of course, I'm happy to change this if people think that would be better.