@stuhlmueller 2016-01-08
This looks like it will make things a lot snappier. Awesome!

If it turns out that running a 700ms computation on page load makes the page less responsive, we should consider delaying it until the first code box is run (or, if we're using Long's new editor, maybe run it in a web worker thread on page load).

Note for posterity: I was initially confused by the use of [`_.isObject`](https://lodash.com/docs#isObject)—I didn't expect this function to treat arrays as objects, too, but it does, and it needs to for `copyAst` to work correctly.